feat: refactor role process
