+++ lib/next-auth/role.ts
@@ -1,0 +1,6 @@
+export const ROLE_USER = 'user';
+export const ROLE_ADMIN = 'admin';
+
+export function checkAdmin(role: string | null | undefined): boolean {
+  return !!role && role === ROLE_ADMIN;
+}
